Add delete guard for report money transfers

+++ src/AppBundle/Controller/Report/MoneyTransferController.php
@@ -216,6 +216,7 @@ class MoneyTransferController extends AbstractController
 
     /**
      * @Route("/report/{reportId}/money-transfers/{transferId}/delete", name="money_transfers_delete")
+     * @Template("AppBundle:Common:confirmDelete.html.twig")
      *
      * @param int $id
      *
@@ -225,14 +226,34 @@ class MoneyTransferController extends AbstractController
     {
         $report = $this->getReportIfNotSubmitted($reportId, self::$jmsGroups);
 
-        $this->getRestClient()->delete("/report/{$reportId}/money-transfers/{$transferId}");
+        $form = $this->createForm(FormDir\ConfirmDeleteType::class);
+        $form->handleRequest($request);
 
-        $request->getSession()->getFlashBag()->add(
-            'notice',
-            'Money transfer deleted'
-        );
+        if ($form->isValid()) {
+            $this->getRestClient()->delete("/report/{$reportId}/money-transfers/{$transferId}");
 
-        return $this->redirect($this->generateUrl('money_transfers_summary', ['reportId' => $reportId]));
+            $request->getSession()->getFlashBag()->add(
+                'notice',
+                'Money transfer deleted'
+            );
+
+            return $this->redirect($this->generateUrl('money_transfers_summary', ['reportId' => $reportId]));
+        }
+
+        $transfer = $report->getMoneyTransferWithId($transferId);
+
+        return [
+            'report' => $report,
+            'translationDomain' => 'report-money-transfer',
+            'subject' => 'transfer',
+            'form' => $form->createView(),
+            'summary' => [
+                ['label' => 'Transferred from', 'value' => $transfer->getAccountFrom()->getNameOneLine()],
+                ['label' => 'Transferred to', 'value' => $transfer->getAccountTo()->getNameOneLine()],
+                ['label' => 'Amount', 'value' => $transfer->getAmount(), 'format' => 'money'],
+            ],
+            'backLink' => $this->generateUrl('money_transfers_summary', ['reportId' => $reportId]),
+        ];
     }
 
     /**
